move expire thread interval into videobridgeconfig
