+++ src/main/resources/reference.conf
@@ -1,6 +1,7 @@
 videobridge {
   # The interval at which the expire thread runs
   expire-thread-interval=1 minute
+  enabled-apis=[]
   health {
     # The interval between health checks
     interval=10 seconds
